# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API to interface with Sidero Omni.",
        "title": "Talos Omni Control API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clusters": {
            "get": {
                "description": "Get a list of all clusters in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "List all clusters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ClusterResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/machines/{id}/status": {
            "get": {
                "description": "Get detailed status information about a specific machine. This endpoint is deprecated - status information is now included in the main machine endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get machine status (deprecated: use GET /machines/{id} instead)",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MachineResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/machinesets": {
            "get": {
                "description": "Get a list of all machine sets in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machinesets"
                ],
                "summary": "List all machine sets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.MachineSetResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/machinesets/{id}/destroy-status": {
            "get": {
                "description": "Get the status of machine set destruction operation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machinesets"
                ],
                "summary": "Get machine set destroy status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine Set ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MachineSetDestroyStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/machinesets/{id}/status": {
            "get": {
                "description": "Get status information for a specific machine set",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machinesets"
                ],
                "summary": "Get machine set status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine Set ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MachineSetStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/ongoingtasks": {
            "get": {
                "description": "Get a list of all currently running tasks in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ongoingtasks"
                ],
                "summary": "List all ongoing tasks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by resource ID",
                        "name": "resource",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.OngoingTaskResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/ongoingtasks/{id}": {
            "get": {
                "description": "Get detailed information about a specific ongoing task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ongoingtasks"
                ],
                "summary": "Get a single ongoing task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ongoing Task ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.OngoingTaskResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ClusterResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "features": {
                    "type": "object",
                    "properties": {
                        "disk_encryption": {
                            "type": "boolean"
                        },
                        "workload_proxy": {
                            "type": "boolean"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "kubernetes_version": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                }
            }
        },
...
        "handlers.MachineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "arch": {
                    "type": "string"
                },
                "connected": {
                    "type": "boolean"
                },
                "hostname": {
                    "description": "Status fields (from MachineStatus resource)",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "last_error": {
                    "type": "string"
                },
                "maintenance": {
                    "type": "boolean"
                },
                "management_address": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                },
                "use_grpc_tunnel": {
                    "type": "boolean"
                }
            }
        },
        "handlers.MachineSetDestroyStatusResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                }
            }
        },
...
        "handlers.MachineSetResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "delete_strategy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "machine_class": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "update_strategy": {
                    "type": "string"
                }
            }
        },
        "handlers.MachineSetStatusResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "config_hash": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "locked_updates": {
                    "type": "integer"
                },
                "machines": {
                    "type": "object",
                    "properties": {
                        "connected": {
                            "type": "integer"
                        },
                        "healthy": {
                            "type": "integer"
                        },
                        "requested": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "namespace": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "ready": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.OngoingTaskResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}